rename TypeSerialization to CborSerialization
