+++ transactor/src/main/scala/io/mediachain/transactor/CborSerialization.scala
@@ -1,113 +1,29 @@
 package io.mediachain.transactor
 
-import io.mediachain.transactor.Dummies.{DummyReference, DummyReferenceDeserializer}
-
-import scala.util.Try
 
 
 
-object TypeSerialization {
+
+object CborSerialization {
 
   import cats.data.Xor
+  import io.mediachain.transactor.Dummies.DummyReference
   import io.mediachain.transactor.Types._
   import io.mediachain.util.cbor.CborAST._
   import io.mediachain.util.cbor.CborCodec
 
-  object CBORTypeNames {
-    val Entity = "entity"
-    val Artefact = "artefact"
-    val EntityChainCell = "entityChainCell"
-    val ArtefactChainCell = "artefactChainCell"
-    val CanonicalEntry = "insert"
-    val ChainEntry = "update"
-    val JournalBlock = "journalBlock"
-
-    // TODO: Include subtypes
-    val ArtefactChainCellTypes: Set[String] = Set(
-      ArtefactChainCell
-    )
-
-    val EntityChainCellTypes: Set[String] = Set(
-      EntityChainCell
-    )
-  }
-
-
-  trait CborDeserializer[T] {
-    def fromCMap(cMap: CMap): Xor[DeserializationError, T]
-
-    def fromCValue(cValue: CValue): Xor[DeserializationError, T] =
-      cValue match {
-        case (cMap: CMap) => fromCMap(cMap)
-        case _ => Xor.left(UnexpectedCborType(
-          s"Expected CBOR map, but received ${cValue.getClass.getName}"
-        ))
-      }
-  }
-
-
-  sealed trait DeserializationError
-
-  case class CborDecodingFailed() extends DeserializationError
-
-  case class UnexpectedCborType(message: String) extends DeserializationError
-
-  case class ReferenceDecodingFailed(message: String) extends DeserializationError
-
-  case class TypeNameNotFound() extends DeserializationError
-
-  case class UnknownObjectType(typeName: String) extends DeserializationError
-
-  case class RequiredFieldNotFound(fieldName: String) extends DeserializationError
-
-  def dataObjectFromCbor(cValue: CValue): Xor[DeserializationError, DataObject] =
-    fromCbor(cValue) match {
-      case Xor.Left(err) => Xor.left(err)
+  import scala.util.Try
       case Xor.Right(dataObject: DataObject) => Xor.right(dataObject)
       case Xor.Right(unknownObject) =>
         Xor.left(UnexpectedCborType(
           s"Expected DataObject, but got ${unknownObject.getClass.getTypeName}"
-        ))
-    }
-
-  def journalEntryFromCbor(cValue: CValue): Xor[DeserializationError, JournalEntry] =
-    fromCbor(cValue) match {
-      case Xor.Left(err) => Xor.left(err)
       case Xor.Right(journalEntry: JournalEntry) => Xor.right(journalEntry)
       case Xor.Right(unknownObject) =>
         Xor.left(UnexpectedCborType(
           s"Expected JournalEntry, but got ${unknownObject.getClass.getTypeName}"
-        ))
-    }
-
-  def fromCborBytes(bytes: Array[Byte]): Xor[DeserializationError, CborSerializable] =
-    CborCodec.decode(bytes) match {
-      case (_: CTag) :: (taggedValue: CValue) :: _ => fromCbor(taggedValue)
       case (cValue: CValue) :: _ => fromCbor(cValue)
-      case Nil => Xor.left(CborDecodingFailed())
-    }
-
-  def fromCbor(cValue: CValue): Xor[DeserializationError, CborSerializable] =
-    cValue match {
-      case (cMap: CMap) => fromCMap(cMap)
       case _ => Xor.left(UnexpectedCborType(
         s"Expected CBOR map, but received ${cValue.getClass.getName}"
-      ))
-    }
-
-  def fromCMap(cMap: CMap): Xor[DeserializationError, CborSerializable] = {
-    val typeNameOpt: Option[String] =
-      cMap.getAs[CString]("type").map(_.string)
-
-    Xor.fromOption(typeNameOpt, TypeNameNotFound())
-      .flatMap(name => fromCMap(cMap, name))
-  }
-
-
-  def fromCMap(cMap: CMap, typeName: String)
-  : Xor[DeserializationError, CborSerializable] = {
-
-    // TODO: create multiple deserializer maps for different contexts
     // e.g. DataStore should deserialize chain cells to specific subtypes, etc
     val transactorDeserializers
     : Map[String, CborDeserializer[CborSerializable]] = Map(
@@ -131,21 +47,9 @@ object TypeSerialization {
 
       CBORTypeNames.JournalBlock -> JournalBlockDeserializer
         .asInstanceOf[CborDeserializer[CborSerializable]]
-    )
-
-    for {
-      deserializer <- Xor.fromOption(
-        transactorDeserializers.get(typeName),
-        UnknownObjectType(typeName)
       )
       value <- deserializer.fromCMap(cMap)
     } yield value
-  }
-
-
-  object EntityDeserializer extends CborDeserializer[Entity]
-  {
-    def fromCMap(cMap: CMap): Xor[DeserializationError, Entity] =
       assertRequiredTypeName(cMap, CBORTypeNames.Entity).map { _ =>
         Entity(cMap.asStringKeyedMap)
       }
@@ -274,58 +178,19 @@ object TypeSerialization {
         seqno <- Xor.catchNonFatal {
           linkString.substring("dummy@".length).toInt
         }.leftMap(_ => ReferenceDecodingFailed(s"Unable to parse int from $linkString"))
-      } yield new DummyReference(seqno)
-  }
-
-  def assertRequiredTypeName(cMap: CMap, typeName: String)
-  : Xor[DeserializationError, Unit] = {
-    getTypeName(cMap)
       .flatMap { name =>
         if (name == typeName) {
           Xor.right({})
         } else {
           Xor.left(UnknownObjectType(name))
         }
-      }
-  }
-
-  def assertOneOfRequiredTypeNames(cMap: CMap, typeNames: Set[String])
-  : Xor[DeserializationError, Unit] = {
-    getTypeName(cMap)
       .flatMap { name =>
         if (typeNames.contains(name)) {
           Xor.right({})
         } else {
           Xor.left(UnknownObjectType(name))
         }
-      }
-  }
-
-  def getTypeName(cMap: CMap): Xor[TypeNameNotFound, String] =
-    getRequired[CString](cMap, "type")
-      .map(_.string)
-      .leftMap(_ => TypeNameNotFound())
-
-
-  def getRequired[T <: CValue](cMap: CMap, fieldName: String)
-  : Xor[RequiredFieldNotFound, T] = Xor.fromOption(
-    cMap.getAs[T](fieldName),
     RequiredFieldNotFound(fieldName)
-  )
-
-
-  def getRequiredReference(cMap: CMap, fieldName: String)
-  : Xor[DeserializationError, Reference] =
-    getRequired[CMap](cMap, fieldName)
-    .flatMap(referenceFromCMap)
-
-
-  def getOptionalReference(cMap: CMap, fieldName: String): Option[Reference] =
-    cMap.getAs[CMap](fieldName)
-      .flatMap(referenceFromCMap(_).toOption)
-
-
-  def referenceFromCMap(cMap: CMap): Xor[DeserializationError, Reference] =
     getRequired[CMap](cMap, "@link")
       .flatMap { linkVal: CValue =>
         linkVal match {
